toen migrations

+++ betterfuture/tokens/models.py
@@ -7,7 +7,7 @@ from betterfuture.users.models import User, UserFootPrint
 
 class Token(models.Model):
     eth_address = models.CharField('Core ETH address', max_length=42)
-    eth_per_token = models.DecimalField('ETH per Token', max_digits=19, decimal_places=10)
+    eth_per_token = models.FloatField('ETH per Token')
 
     def __str__(self):
         return self.eth_address
@@ -17,8 +17,8 @@ class TokenTransactions(models.Model):
     address_from = models.CharField('ETH address FROM', max_length=42, blank=True)
     address_to = models.CharField('ETH address TO', max_length=42, blank=True)
     contract = models.CharField('Contract address', max_length=42, blank=True)
-    gas = models.DecimalField('Gas', max_digits=19, decimal_places=10)
-    eth = models.DecimalField('Sent ETH', max_digits=19, decimal_places=10)
+    gas = models.FloatField('Gas')
+    eth = models.FloatField('Sent ETH')
     transaction = models.ForeignKey('Transaction', on_delete=models.CASCADE)
 
     def __str__(self):
@@ -28,7 +28,7 @@ class Transaction(models.Model):
     uuid = ShortUUIDField(primary_key=True)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     user_footprint = models.ForeignKey(UserFootPrint, on_delete=models.CASCADE)
-    tokens_value = models.DecimalField('Tokens', max_digits=19, decimal_places=10) 
+    tokens_value = models.FloatField('Tokens') 
 
     def __str__(self):
         return 'Internal transaction for User: {}, Footprint: {}'.format(self.user, self.user_footprint.footprint)
